Lab 6
One of the most common applications of stack is in web browsers. To implement the undo / 
redo functionality a stack data structure is used. In this section you will create a web browser 
with that functionality.
You are free to design the web browser class as you wish (give it a unique name). However, 
follow the guidelines:
Restrictions: [1 mark]
 You can only use objects of your stack class as data structure (No Arrays, List, Queues, 
and Trees etc). 
 You may create variables if you wish.
 For filing you can use string from std.
Requirements:
1. Prompt the user to:
 give a URL: (e.g www.google.com for simplicity you can use “google” “facebook”): 
[2 marks]
 Open a file named “google.txt”, “facebook.txt” and display the content of file on 
screen
 if file doesn’t exist, error: “Invalid URL”
 Undo: if a URL was opened after another URL the previous goes into undo [1 mark]
 Open and display content of previous URL, do nothing if no previous URL exists.
 Redo: if an undo was called from a URL, then that URL goes into redo [1 mark]
 if a redo URL exist, open and display content on screen, else do nothing
 History: [2 marks]
 Search for the URL in undo stack. print “You have visited/not visited this website” You can only use stack object and its operations to search.
 Make sure the undo stack remains intact
 (Optional) you can search in redo stack as well. [bonus 1 mark]
 (Optional) you can permanently store all history in a file (Data persistence) so 
when the program launch you can search in undo, redo stacks as well as past 
history if you like. [bonus 1 mark]
Try to make re-usable methods. For example if you create a method: display(filename){} in 
your class, You can use this in giving a URL, Undo and Redo methods